revert estimator
