+++ quri_algo/algo/phase_estimation/interface.py
@@ -15,7 +15,7 @@ from typing import Any, Generic, Optional, Sequence
 import numpy as np
 from quri_parts.core.estimator import Estimate, Estimates
 
-from quri_algo.algo.estimator import StateT
+from quri_algo.core.estimator import StateT
 
 
 @dataclass
