+++ quri_algo/algo/phase_estimation/spe/utils/exp_val_collector.py
@@ -13,7 +13,7 @@ from typing import Callable, Generic, NamedTuple, Optional, Sequence
 from quri_parts.core.estimator import Estimate
 from typing_extensions import TypeAlias
 
-from quri_algo.algo.estimator import StateT
+from quri_algo.core.estimator import StateT
 
 from .fourier.coefficient import SPEFourierCoefficient
