+++ quri_algo/core/estimator/hadamard_test.py
