+++ tests/algo/spe/test_gaussian.py
@@ -17,12 +17,12 @@ from quri_parts.core.operator import PAULI_IDENTITY, Operator
 from quri_parts.core.state import CircuitQuantumState, quantum_state
 
 import quri_algo.algo.phase_estimation.spe as spe
-from quri_algo.algo.estimator import OperatorPowerEstimatorBase
-from quri_algo.algo.estimator.time_evolution import (
-    TimeEvolutionExpectationValueEstimator,
-)
 from quri_algo.algo.phase_estimation.spe.gaussian import (
     get_recommended_gaussian_parameter,
+)
+from quri_algo.core.estimator import OperatorPowerEstimatorBase
+from quri_algo.core.estimator.time_evolution import (
+    TimeEvolutionExpectationValueEstimator,
 )
 from quri_algo.problem import QubitHamiltonianInput
