+++ tests/algo/spe/test_lt22.py
@@ -16,11 +16,11 @@ import numpy as np
 from quri_parts.core.state import CircuitQuantumState, quantum_state
 
 import quri_algo.algo.phase_estimation.spe as spe
-from quri_algo.algo.estimator import OperatorPowerEstimatorBase
-from quri_algo.algo.estimator.time_evolution import (
+from quri_algo.algo.phase_estimation.spe import SingleSignalLT22PhaseEstimation
+from quri_algo.core.estimator import OperatorPowerEstimatorBase
+from quri_algo.core.estimator.time_evolution import (
     TimeEvolutionExpectationValueEstimator,
 )
-from quri_algo.algo.phase_estimation.spe import SingleSignalLT22PhaseEstimation
 
 
 class _Estimate(NamedTuple):
