+++ tests/algo/spe/test_signal.py
@@ -18,7 +18,6 @@ import numpy.typing as npt
 from quri_parts.core.state import CircuitQuantumState, quantum_state
 
 import quri_algo.algo.phase_estimation.spe.utils.fourier as fourier
-from quri_algo.algo.estimator import State
 from quri_algo.algo.phase_estimation.spe.utils.exp_val_collector import (
     ExpectationValueCollector,
     SPESample,
@@ -29,6 +28,7 @@ from quri_algo.algo.phase_estimation.spe.utils.signal import (
     SPEDiscreteSignalFunction,
     StepFunctionSignalGenerator,
 )
+from quri_algo.core.estimator import State
 
 
 class _Estimate(NamedTuple):
