+++ tests/core/estimators/test_hadamard_test.py
@@ -22,13 +22,13 @@ from quri_parts.core.state import CircuitQuantumState, quantum_state
 from quri_parts.qulacs.sampler import create_qulacs_vector_ideal_sampler
 from scipy.linalg import expm
 
-from quri_algo.algo.estimator.hadamard_test import (
+from quri_algo.circuit.interface import CircuitFactory
+from quri_algo.core.estimator.hadamard_test import (
     HadamardTest,
     get_hadamard_test_ancilla_qubit_counter,
     remap_state_for_hadamard_test,
     shift_state_circuit,
 )
-from quri_algo.circuit.interface import CircuitFactory
 from quri_algo.problem import Problem
