+++ tests/core/estimators/time_evo_estimators/test_exact_spectrum_estimator.py
@@ -12,7 +12,7 @@ import numpy as np
 from quri_parts.core.operator import Operator, get_sparse_matrix, pauli_label
 from quri_parts.core.state import quantum_state
 
-from quri_algo.algo.estimator.time_evolution.exact_spectrum import (
+from quri_algo.core.estimator.time_evolution.exact_spectrum import (
     ExactTimeEvolutionExpectationValueEstimator,
 )
 from quri_algo.problem.operators.hamiltonian import QubitHamiltonianInput
